+++ test/test_byte_array.rb
@@ -37,4 +37,13 @@ class TestByteArray < Test::Unit::TestCase
     assert_equal 1, hash[ HBase::ByteArray.new("Hello") ]
     assert_equal 2, hash[ HBase::ByteArray.new("World".to_java_bytes) ]
   end
+
+  def test_concat
+    concat = HBase::ByteArray(100) + HBase::ByteArray(200)
+    assert_instance_of HBase::ByteArray, concat
+    assert_equal 16, concat.to_java_bytes.to_a.length
+
+    assert_equal 100, HBase::Util.from_bytes( :fixnum, concat.to_java_bytes.to_a[0, 8].to_java(Java::byte) )
+    assert_equal 200, HBase::Util.from_bytes( :fixnum, concat.java.to_a[8, 8].to_java(Java::byte) )
+  end
 end